+++ packages/alfa-parser/src/parser.ts
@@ -4,7 +4,8 @@ import { None, Option } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Result, Err } from "@siteimprove/alfa-result";
 import { Refinement } from "@siteimprove/alfa-refinement";
-import { Thunk } from "@siteimprove/alfa-thunk";
+
+const { not } = Predicate;
 
 export type Parser<I, T, E = never, A extends Array<unknown> = []> = (
   input: I,
@@ -51,57 +52,55 @@ export namespace Parser {
   export function filter<I, T, U extends T, E, A extends Array<unknown> = []>(
     parser: Parser<I, T, E, A>,
     refinement: Refinement<T, U>,
-    ifError: Thunk<E>
+    ifError: Mapper<T, E>
   ): Parser<I, U, E, A>;
 
   export function filter<I, T, E, A extends Array<unknown> = []>(
     parser: Parser<I, T, E, A>,
     predicate: Predicate<T>,
-    ifError: Thunk<E>
+    ifError: Mapper<T, E>
   ): Parser<I, T, E, A>;
 
   export function filter<I, T, E, A extends Array<unknown> = []>(
     parser: Parser<I, T, E, A>,
     predicate: Predicate<T>,
-    ifError: Thunk<E>
+    ifError: Mapper<T, E>
   ): Parser<I, T, E, A> {
     return mapResult(parser, (value) =>
-      predicate(value) ? Result.of(value) : Err.of(ifError())
+      predicate(value) ? Result.of(value) : Err.of(ifError(value))
     );
+  }
+
+  export function reject<I, T, U extends T, E, A extends Array<unknown> = []>(
+    parser: Parser<I, T, E, A>,
+    refinement: Refinement<T, U>,
+    ifError: Mapper<T, E>
+  ): Parser<I, Exclude<T, U>, E, A>;
+
+  export function reject<I, T, E, A extends Array<unknown> = []>(
+    parser: Parser<I, T, E, A>,
+    predicate: Predicate<T>,
+    ifError: Mapper<T, E>
+  ): Parser<I, T, E, A>;
+
+  export function reject<I, T, E, A extends Array<unknown> = []>(
+    parser: Parser<I, T, E, A>,
+    predicate: Predicate<T>,
+    ifError: Mapper<T, E>
+  ): Parser<I, T, E, A> {
+    return filter(parser, not(predicate), ifError);
   }
 
   export function zeroOrMore<I, T, E, A extends Array<unknown> = []>(
     parser: Parser<I, T, E, A>
   ): Parser<I, Array<T>, E, A> {
-    return (input, ...args) => {
-      const values: Array<T> = [];
-
-      let value: T;
-
-      while (true) {
-        const result = parser(input, ...args);
-
-        if (result.isOk()) {
-          [input, value] = result.get();
-          values.push(value);
-        } else {
-          break;
-        }
-      }
-
-      return Result.of([input, values]);
-    };
+    return takeAtLeast(parser, 0);
   }
 
   export function oneOrMore<I, T, E, A extends Array<unknown> = []>(
     parser: Parser<I, T, E, A>
   ): Parser<I, Array<T>, E, A> {
-    return flatMap(parser, (head) =>
-      map(zeroOrMore(parser), (tail) => {
-        tail.unshift(head);
-        return tail;
-      })
-    );
+    return takeAtLeast(parser, 1);
   }
 
   export function take<I, T, E, A extends Array<unknown> = []>(
@@ -279,6 +278,17 @@ export namespace Parser {
   }
 
   export function delimited<I, T, E, A extends Array<unknown> = []>(
+    parser: Parser<I, unknown, E, A>,
+    separator: Parser<I, T, E, A>
+  ): Parser<I, T, E, A>;
+
+  export function delimited<I, T, E, A extends Array<unknown> = []>(
+    left: Parser<I, unknown, E, A>,
+    separator: Parser<I, T, E, A>,
+    right: Parser<I, unknown, E, A>
+  ): Parser<I, T, E, A>;
+
+  export function delimited<I, T, E, A extends Array<unknown> = []>(
     left: Parser<I, unknown, E, A>,
     separator: Parser<I, T, E, A>,
     right: Parser<I, unknown, E, A> = left
@@ -289,10 +299,21 @@ export namespace Parser {
   }
 
   export function separated<I, T, U, E, A extends Array<unknown> = []>(
+    parser: Parser<I, T, E, A>,
+    separator: Parser<I, unknown, E, A>
+  ): Parser<I, [T, T], E, A>;
+
+  export function separated<I, T, U, E, A extends Array<unknown> = []>(
     left: Parser<I, T, E, A>,
     separator: Parser<I, unknown, E, A>,
     right: Parser<I, U, E, A>
-  ): Parser<I, [T, U], E, A> {
+  ): Parser<I, [T, U], E, A>;
+
+  export function separated<I, T, E, A extends Array<unknown> = []>(
+    left: Parser<I, T, E, A>,
+    separator: Parser<I, unknown, E, A>,
+    right: Parser<I, T, E, A> = left
+  ): Parser<I, [T, T], E, A> {
     return flatMap(left, (left) =>
       flatMap(separator, () => map(right, (right) => [left, right]))
     );
@@ -311,7 +332,7 @@ export namespace Parser {
     );
   }
 
-  export function eof<I extends Iterable<unknown>, E>(
+  export function end<I extends Iterable<unknown>, E>(
     ifError: Mapper<I extends Iterable<infer T> ? T : unknown, E>
   ): Parser<I, void, E> {
     return (input) => {
@@ -322,4 +343,9 @@ export namespace Parser {
       return Result.of([input, undefined]);
     };
   }
+
+  /**
+   * @deprecated Use `end()`
+   */
+  export const eof = end;
 }